pyproject.toml
[build-system]
requires = ["setuptools", "cython"]
build-backend = "setuptools.build_meta"

[project]
name = "nefelis"
version = "0.1"
authors = [ { name="Rémy O", email="remyoudompheng@gmail.com" } ]
description = "NFS implementation using GPU"
readme = "README.md"
requires-python = ">=3.11"
license = "BSD-3-Clause"
license-files = ["LICENSE"]

dependencies = [
  "python-flint>=0.7",
  "kp==0.9.0",
  "tqdm>=4.0",
  "networkx>=3.2"
]

[project.scripts]
nefelis = "nefelis:main"

[tool.setuptools]
ext-modules = [{name = "nefelis_flint_extras", sources = ["nefelis/nefelis_flint_extras.pyx"], libraries = ["flint"]}]

[tool.setuptools.packages.find]
include = ['nefelis*']

[tool.setuptools.package-data]
nefelis = ["vulkan/*.comp"]

[tool.pytest.ini_options]
addopts = "--doctest-modules"
pythonpath = ["."]

[tool.ruff.lint]
# It's okay to name variables 'l'
ignore = ["E741"]
